Move index-server inside src/server folder
