+++ apps/front/package.json
@@ -8,7 +8,7 @@
   "scripts": {
     "dev": "node server.js",
     "build:client": "vite build --outDir dist/client",
-    "build:server": "vite build --ssr src/index-server.tsx --outDir dist/server",
+    "build:server": "vite build --ssr src/server/index-server.tsx --outDir dist/server",
     "build:scripts": "vite build -c vite.scripts.config.ts",
     "build:static": "vite build --outDir dist/static",
     "build": "npm run build:server && npm run build:scripts && npm run build:static && npm run generate",
