+++ apps/front/prerender/prerender.ts
@@ -1,7 +1,7 @@
 // @ts-ignore
 import path from "path"
 // @ts-ignore
-import { render } from "~/index-server"
+import { render } from "~/server/index-server"
 import * as mfs from "../config/helpers/mfs.js"
 import config from "../config/config.js"
 import palette from "../config/helpers/palette.js"
