+++ apps/front/src/server/index-server.tsx
@@ -1,13 +1,13 @@
 import fetch from "cross-fetch"
 import * as React from "react"
-import { routes } from "./routes"
-import App from "./components/app/App"
-import { languages, showDefaultLangInUrl } from "./languages"
+import { routes } from "~/routes"
+import App from "../components/app/App"
+import { languages, showDefaultLangInUrl } from "~/languages"
 import { LangService, requestStaticPropsFromRoute, Router } from "@cher-ami/router"
-import { GlobalDataContext } from "./store/GlobalDataContext"
-import { preventSlashes } from "../config/helpers/prevent-slashes.js"
+import { GlobalDataContext } from "~/store/GlobalDataContext"
+import { preventSlashes } from "../../config/helpers/prevent-slashes.js"
 import { loadEnv } from "vite"
-import { TScriptsObj } from "../prerender/helpers/ManifestParser"
+import { TScriptsObj } from "../../prerender/helpers/ManifestParser"
 import { CherScripts } from "~/server/helpers/CherScripts"
 import { RawScript } from "~/server/helpers/RawScript"
 import { ViteDevScripts } from "~/server/helpers/ViteDevScripts"
