+++ docs/02.vite-configuration.md
@@ -15,5 +15,5 @@ Vite's configuration is managed by two main files:
 
 Two entry points are set:
 
-- server side [src/index-server.tsx](../src/index-server.tsx)
+- server side [src/index-server.tsx](../src/server/index-server.tsx)
 - client side [src/index-client.tsx](../src/index.tsx)
